Updated setup views for add custom use case
